tests/conformance/requirements.txt
PyYAML>=6.0.0
jsonschema>=4.18.0
